requirements.txt
fastapi==0.109.1
uvicorn==0.25.0
python-multipart==0.0.6
langchain==0.1.12
langchain-community
langchain-groq
groq==0.4.1
pypdf==4.0.1
python-dotenv==1.0.0
pydantic==2.6.3
faiss-cpu==1.7.4
langchain-text-splitters==0.0.1